+++ CHANGES.rst
@@ -15,7 +15,6 @@ Changelog
 
 **Fixed**
 - #991 New client contacts do not have access to their own AR Templates
-
 - #996 Hide checkbox labels on category expansion
 - #990 Fix client analysisspecs view
